Move external,noderefutil,remote out of pkg/controller
