+++ cmd/clusterctl/clusterdeployer/bootstrap/minikube/minikube.go
@@ -25,7 +25,7 @@ import (
 
 	"github.com/pkg/errors"
 	"k8s.io/klog"
-	"sigs.k8s.io/cluster-api/pkg/util"
+	"sigs.k8s.io/cluster-api/util"
 )
 
 const (
