+++ cmd/clusterctl/cmd/alpha_phase_apply_machines.go
@@ -24,7 +24,7 @@ import (
 	"k8s.io/klog"
 	"sigs.k8s.io/cluster-api/cmd/clusterctl/clusterdeployer/clusterclient"
 	"sigs.k8s.io/cluster-api/cmd/clusterctl/phases"
-	"sigs.k8s.io/cluster-api/pkg/util"
+	"sigs.k8s.io/cluster-api/util"
 )
 
 type AlphaPhaseApplyMachinesOptions struct {
