+++ cmd/clusterctl/phases/getkubeconfig.go
@@ -25,7 +25,7 @@ import (
 
 	"k8s.io/klog"
 	"sigs.k8s.io/cluster-api/cmd/clusterctl/clusterdeployer/clusterclient"
-	"sigs.k8s.io/cluster-api/pkg/util"
+	"sigs.k8s.io/cluster-api/util"
 )
 
 const (
