+++ hack/tools.go
@@ -1,3 +1,5 @@
+// +build tools
+
 /*
 Copyright 2019 The Kubernetes Authors.
