+++ test/integration/cluster/cluster_test.go
@@ -1,3 +1,5 @@
+// +build e2e
+
 /*
 Copyright 2019 The Kubernetes Authors.
